# Changes custom_attribute virtual_attributes to support AREL/SQL