@NickLaMuro 2018-06-20
@jrafanie yeah, this gets exercised with some of the tests here:

https://github.com/ManageIQ/manageiq/blob/362b5ff7d440a3a9efa6c1ced867c4f78b9a9458/spec/models/miq_report_spec.rb#L100-L110

These were the tests that were failing when I mentioned them failing on CI here:  https://github.com/ManageIQ/manageiq/pull/17615#issuecomment-398627754