@NickLaMuro 2018-06-20
@kbrock mind weighing in on this PR?