@simaishi 2018-06-26
Backported to Gaprindashvili via https://github.com/ManageIQ/manageiq/pull/17629