@simaishi 2018-06-26
Backported https://github.com/ManageIQ/manageiq/pull/17629 to Fine branch.